refactor: move alt header, footer, and sections to their own components
